// src/types/internal/permissions/PermissionCheckStatus.ts
// eslint-disable-next-line import/prefer-default-export
export enum PermissionCheckStatus {
  NOT_PERMITTED = 'NOT_PERMITTED', // Generic not permitted status - default
  NOT_IN_CASELOAD = 'NOT_IN_CASELOAD',
  NOT_ACTIVE_CASELOAD = 'NOT_ACTIVE_CASELOAD',
  RESTRICTED_PATIENT = 'RESTRICTED_PATIENT',
  PRISONER_IS_RELEASED = 'PRISONER_IS_RELEASED',
  PRISONER_IS_TRANSFERRING = 'PRISONER_IS_TRANSFERRING',
  EXCEEDS_TIME_RESTRICTION = 'EXCEEDS_TIME_RESTRICTION',
  ROLE_NOT_PRESENT = 'ROLE_NOT_PRESENT',
  NOT_PRISON_USER = 'NOT_PRISON_USER',
  OK = 'OK',
}